chore: slight change of structure to aid discussion
